refactor register form into separate component wrapped with suspense
